Consolidate code used by groups tiles
Move everything tile-related and used only by the groups page to
src/components/Groups/Tiles.
