+++ static-site/src/components/Groups/Tiles/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import * as Styles from "./styles";
-import { H1 } from "../splash/styles";
-import { MediumSpacer } from "../../layouts/generalComponents";
+import { H1 } from "../../splash/styles";
+import { MediumSpacer } from "../../../layouts/generalComponents";
 import Padlock from "./padlock";
 
 class Tiles extends React.Component {
@@ -20,7 +20,7 @@ class Tiles extends React.Component {
                 </Styles.BottomRightLabel>
               ) : null}
               {/* eslint-disable-next-line @typescript-eslint/no-var-requires */}
-              {d.img ? <Styles.TileImg src={require(`../../../static/splash_images/${d.img}`).default.src} alt={""} color={d.color}/> : null}
+              {d.img ? <Styles.TileImg src={require(`../../../../static/splash_images/${d.img}`).default.src} alt={""} color={d.color}/> : null}
             </a>
           </Styles.TileInner>
         </Styles.TileOuter>
