+++ static-site/src/components/Groups/Tiles/styles.tsx
